This PR addresses issue https://github.com/scikit-learn-contrib/scikit-dimension/issues/19
I used this test code to demonstrate the issue and benchmark the solution:
```
import numpy as np
import skdim
import time

np.random.seed(0)
X = np.random.random((20000, 10))

estimator = skdim.id.ESS() # LocalEstimator
n_jobs = 1

t0 = time.time()
lid = skdim.asPointwise(X, estimator, n_neighbors=10, n_jobs=n_jobs)
t1 = time.time()

print(lid.mean())
print(t1-t0)
```
On master, commit b9e88451, by only changing n_jobs I observe the following using a computer with 40 cpu. Note that the timing includes knn computation which is affected by n_jobs, but for these settings knn takes only a couple seconds for n_jobs=1.
| n_jobs | Time (s) | Mean LID | Notes |
| --- | --- | --- | --- |
| 1 | 222.62 | 8.745108 | About 50% of cpu being used at 100% (others at 0%) |
| 2 | - | - | Does not complete. Makes no progress. 100% of cpus at 0% usage |

After the PR changes I observe

| n_jobs | Time (s) | Mean LID | Notes |
| --- | --- | --- | --- |
| 1 | 219.36 | 8.745108 | About 50% of cpu being used at 90% |
| 2 | 208.17 | 8.745108 | About 100% of cpu being used at 90% |
| 20 | 28.47 | 8.745108 | About 100% of cpu being used at 100% |
| 40 | 16.95 | 8.745108 | About 100% of cpu being used at 100% |

Conclusions:
Multiprocessing is now working properly and there is a significant speedup from increasing n_jobs (including both knn computation, and estimator computation). The computed LID values are unchanged.
The ESS estimator code apparently uses more than one cpu even when n_jobs=1 , although it is not doing so efficiently. I have not tried to isolate which step is trying to use multiple cpus.