@JesseCresswell 2024-03-11
Yes fair point for anyone else reading this. The built-in method `ESS.fit()` appears to be more efficient than using `asPointwise`. My example was just for benchmarking.